requirements.txt
tensorflow
tensorflow-datasets
numpy
pillow
streamlit
matplotlib
